Accidentally deleted files, to be re-implemented.
It is a sad story. When I was using neo-tree to move files, I deleted all my source code
in backtrack/ and branch-and-bound/ due to somehow mistyping the new pathname. The only
thing I remember is that, the pathname contains within it a `..`. Neither can I restore
the files nor reproduce the disastrous operation.